# legacy-codebase/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: MyDreamCampus API
  description: |
    Üniversite yönetim sistemi API dokümantasyonu.

    Tüm API'ler Traefik gateway üzerinden `http://localhost/api/...` şeklinde erişilir.

    ## Kimlik Doğrulama
    - `POST /api/auth/login` ile JWT token alınır
    - Token `Authorization: Bearer <token>` header'ı ile gönderilir
    - Traefik forward-auth middleware'i token'ı doğrular ve `X-User-ID`, `X-User-Role`, `X-User-Department` header'larını backend servislere iletir
    - Refresh token HttpOnly cookie olarak saklanır

    ## Roller
    - **admin**: Tam yetki
    - **teacher**: Ders, yoklama, not yönetimi
    - **student**: Kayıt, yoklama, not görüntüleme

    ## Rate Limiting
    Tüm endpoint'ler rate limiting ile korunur. Yanıt header'ları:
    - `X-RateLimit-Limit`: Zaman penceresi başına izin verilen maksimum istek sayısı
    - `X-RateLimit-Remaining`: Kalan istek sayısı
    - `X-RateLimit-Reset`: Limit sıfırlanma zamanı (Unix epoch saniye)

    Limit aşıldığında `429 Too Many Requests` döner.

    ## Hata Yönetimi
    Tüm endpoint'ler aşağıdaki genel hata yanıtlarını dönebilir:
    - `401 Unauthorized`: Geçersiz veya eksik JWT token
    - `429 Too Many Requests`: Rate limit aşıldı
    - `500 Internal Server Error`: Beklenmeyen sunucu hatası

    Hata yanıtları her zaman `ErrorResponse` formatındadır.

    ## Sağlık Kontrolü
    Her servis `GET /health` endpoint'i sunar. Bu endpoint'ler kimlik doğrulama gerektirmez.
  version: 1.0.0
  contact:
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /api/auth/refresh:
    post:
      operationId: refreshToken
      tags: [Auth]
      summary: Token yenile
      description: HttpOnly cookie'deki refresh token ile yeni access token alır.
      security: []
      responses:
        "200":
          description: Yeni access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/auth/logout-all:
    post:
      operationId: logoutAll
      tags: [Auth]
      summary: Tüm oturumlardan çıkış
      description: Kullanıcının tüm aktif oturumlarını sonlandırır.
      responses:
        "200":
          description: Tüm oturumlar sonlandırıldı
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/change-password:
    post:
      operationId: changePassword
      tags: [Auth]
      summary: Şifre değiştir
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Şifre değiştirildi, yeni token döner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangePasswordResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/auth/sessions/{id}:
    delete:
      operationId: deleteSession
      tags: [Auth]
      summary: Belirli bir oturumu sonlandır
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Oturum sonlandırıldı
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token. `POST /api/auth/login` ile alınır."
...
  responses:
    BadRequest:
      description: Geçersiz istek
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Kimlik doğrulama gerekli
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Kayıt bulunamadı
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: "VALIDATION_ERROR"
        error:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "email alanı gereklidir"
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: "işlem başarılı"
...
    RefreshResponse:
      type: object
      properties:
        access_token:
          type: string
        expires_in:
          type: integer
    ChangePasswordRequest:
      type: object
      required: [old_password, new_password]
      properties:
        old_password:
          type: string
          minLength: 8
        new_password:
          type: string
          minLength: 8
    ChangePasswordResponse:
      type: object
      properties:
        message:
          type: string
        access_token:
          type: string
        expires_in:
          type: integer
...